refactor: split analysis/incremental.py into focused modules
Extract change detection, dependency analysis, and carry-forward logic
into _incr_change_detection.py, _incr_dependency.py, and
_incr_carry_forward.py. Keep incremental.py as entry point with
re-exports and classification logic.
